+++ scripts/test.php
@@ -7,6 +7,34 @@ require ('../www/GO.php');
 
 go()->setAuthState(new \go\core\auth\TemporaryState(1));
 
+//go()->getDebugger()->enable(true);
+//go()->getDebugger()->output = true;
+
+$ids = go()->getDbConnection()
+	->select('id')
+	->from('core_user')
+	->orderBy(['lastLogin'=>'desc'])
+	->limit(800)
+	->where('id > 1')
+	->all();
+
+$i = 0;
+while(true) {
+	echo $i++;
+	echo "\n";
+	go()->getDbConnection()->beginTransaction();
+	\go\core\model\User::delete(['id' => $ids]);
+	go()->getDbConnection()->rollBack();
+	\go\core\orm\EntityType::push();
+}
+
+foreach($ids as $id) {
+	\go\core\model\User::delete(['id' => $id]);
+}
+
+
+echo "Done\n";
+
 ini_set("memory_limit", "256M");
 //ini_set("max_execution_time", 20);
 
@@ -21,20 +49,16 @@ for($i = 0; $i<1000; $i++) {
 		throw new \go\core\orm\exception\SaveException($user);
 	}
 
-	$ids[] = $user->id;
+//	$ids[] = $user->id;
+
+	\go\core\model\User::delete(['id' => $user->id]);
 
 //	\go\core\orm\EntityType::push();
 
 }
 
 
-echo "Deleting the users now\n";
-
-\go\core\model\User::delete(['id' => $ids]);
-
-
-
-
+echo "Done\n";
 
 
 //$fakeChanges = [];